local laptop changes merged for multiobject
