+++ Makefile
@@ -1,6 +1,6 @@
 NAME=engine
 CXX=c++
-LIBS=libs
+LIBS=/home/dhorvath/.libs
 CXXFLAGS= -std=c++17 -Ofast -DDEBUG -g -I$(LIBS)/stb
 LDFLAGS = -lglfw -lvulkan -ldl -lpthread -lX11 -lXxf86vm -lXrandr -lXi
