--- render/.gitignore
@@ -1,5 +1,0 @@
-.vscode
-.vscode/*
-obj/
-triangle
-shaders/*.spv
