--- render/shaders/shader.frag
@@ -1,13 +1,0 @@
-#version 450
-
-layout(location = 0) in vec3 fragColor;
-layout(location = 1) in vec2 fragTexCoord;
-
-layout(location = 0) out vec4 outColor;
-
-layout(binding = 1) uniform sampler2D texSampler;
-
-void main() {
-    outColor = texture(texSampler, fragTexCoord);
-    // outColor = vec4(fragTexCoord, 0.0f, 1.0f);
-}
