--- render/src/Model.hpp
@@ -1,71 +1,0 @@
-#ifndef MODEL_HPP
-#define MODEL_HPP
-
-#define GLFW_INCLUDE_VULKAN
-#include <GLFW/glfw3.h>
-
-#define GLM_FORCE_RADIANS
-#define GLM_FORCE_DEPTH_ZERO_TO_ONE
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-
-#include <array>
-#include <vector>
-
-#include "buffer.hpp"
-
-struct Vertex
-{
-    glm::vec3 pos;
-    glm::vec3 color;
-    glm::vec2 texCoord;
-
-    static VkVertexInputBindingDescription getBindingDescription()
-    {
-        VkVertexInputBindingDescription bindingDescription{};
-        bindingDescription.binding = 0;
-        bindingDescription.inputRate = VK_VERTEX_INPUT_RATE_VERTEX;
-
-        bindingDescription.stride = sizeof(Vertex);
-        return bindingDescription;
-    };
-    static std::array<VkVertexInputAttributeDescription, 3> getAttributeDescriptions()
-    {
-        std::array<VkVertexInputAttributeDescription, 3> attributeDescriptions{};
-        attributeDescriptions[0].binding = 0;
-        attributeDescriptions[0].location = 0;
-        attributeDescriptions[0].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[0].offset = offsetof(Vertex, pos);
-
-        attributeDescriptions[1].binding = 0;
-        attributeDescriptions[1].location = 1;
-        attributeDescriptions[1].format = VK_FORMAT_R32G32B32_SFLOAT;
-        attributeDescriptions[1].offset = offsetof(Vertex, color);
-
-        attributeDescriptions[2].binding = 0;
-        attributeDescriptions[2].location = 2;
-        attributeDescriptions[2].format = VK_FORMAT_R32G32_SFLOAT;
-        attributeDescriptions[2].offset = offsetof(Vertex, texCoord);
-
-        return attributeDescriptions;
-    };
-};
-
-class Model
-{
-private:
-    std::string name;
-public:
-    std::vector<Vertex> vertices;
-    std::vector<uint32_t> indices;
-    Buffer vertexBuffer;
-    Buffer indexBuffer;
-    Model(std::string model);
-    void loadModel();
-    void makeVertexBuffer();
-    void makeIndexBuffer();
-    void init();
-    void render(VkCommandBuffer commandbuffer);
-};
-
-#endif
