--- render/src/UniformBufferObject.hpp
@@ -1,16 +1,0 @@
-#ifndef UBO_HPP
-#define UBO_HPP
-
-#define GLM_FORCE_RADIANS
-#define GLM_FORCE_DEPTH_ZERO_TO_ONE
-#include <glm/glm.hpp>
-#include <glm/gtc/matrix_transform.hpp>
-
-struct UniformBufferObject
-{
-    glm::mat4 model;
-    glm::mat4 view;
-    glm::mat4 proj;
-};
-
-#endif
