--- render/src/syncobjects.cpp
@@ -1,23 +1,0 @@
-#include "syncobjects.hpp"
-#include "Vulkan.hpp"
-
-void Syncobjects::makeSyncObjects()
-{
-    imageDoneSemaphores.resize(MAX_FRAMES_IN_FLIGHT);
-    renderFinishedSemaphores.resize(MAX_FRAMES_IN_FLIGHT);
-    inFlightFences.resize(MAX_FRAMES_IN_FLIGHT);
-    VkSemaphoreCreateInfo semaphoreCreateInfo{};
-    semaphoreCreateInfo.sType = VK_STRUCTURE_TYPE_SEMAPHORE_CREATE_INFO;
-    for (int i = 0; i < MAX_FRAMES_IN_FLIGHT; i++)
-    {
-        if (vkCreateSemaphore(VulkanInstance::device, &semaphoreCreateInfo, nullptr, &imageDoneSemaphores[i]) != VK_SUCCESS)
-            throw std::runtime_error("Failed to create semaphore");
-        if (vkCreateSemaphore(VulkanInstance::device, &semaphoreCreateInfo, nullptr, &renderFinishedSemaphores[i]) != VK_SUCCESS)
-            throw std::runtime_error("Failed to create semaphore");
-        VkFenceCreateInfo fenceCreateInfo{};
-        fenceCreateInfo.sType = VK_STRUCTURE_TYPE_FENCE_CREATE_INFO;
-        fenceCreateInfo.flags = VK_FENCE_CREATE_SIGNALED_BIT;
-        if (vkCreateFence(VulkanInstance::device, &fenceCreateInfo, nullptr, &inFlightFences[i]) != VK_SUCCESS)
-            throw std::runtime_error("Failed to create fence");
-    }
-}
