--- render/src/window.cpp
@@ -1,18 +1,0 @@
-#include "window.hpp"
-#include <iostream>
-#include "app.hpp"
-void Window::init()
-{
-    glfwInit();
-    glfwWindowHint(GLFW_RESIZABLE, GLFW_TRUE);
-    glfwWindowHint(GLFW_CLIENT_API, GLFW_NO_API);
-    win = glfwCreateWindow(WIDTH, HEIGHT, "FUNny", nullptr, nullptr);
-    glfwSetWindowUserPointer(win, this);
-    glfwSetFramebufferSizeCallback(win, framebufferResizeCallback);
-}
-
-static void framebufferResizeCallback(GLFWwindow *win, int height, int width)
-{
-    auto app = reinterpret_cast<App *>(glfwGetWindowUserPointer(win));
-    app->frameResize = true;
-}
