+++ resources/shaders/shader.vert
@@ -6,6 +6,10 @@ layout(binding = 0) uniform UniformBufferObject {
     mat4 proj;
 } ubo;
 
+layout(push_constant) uniform push_constant {
+    mat4 model;
+}ps;
+
 layout(location = 0) in vec3 inPosition;
 layout(location = 1) in vec3 inColor;
 layout(location = 2) in vec2 texCoord;
@@ -14,7 +18,7 @@ layout(location = 0) out vec3 fragColor;
 layout(location = 1) out vec2 fragTexCoord;
 
 void main() {
-    gl_Position = ubo.proj * ubo.view * ubo.model * vec4(inPosition, 1.0);
+    gl_Position = ubo.proj * ubo.view * ps.model * ubo.model * vec4(inPosition, 1.0);
     fragColor = inColor;
     fragTexCoord = texCoord;
 }
