+++ src/render/Model.hpp
@@ -12,6 +12,8 @@
 #include <array>
 #include <vector>
 
+#include "buffer.hpp"
+
 struct Vertex
 {
     glm::vec3 pos;
@@ -51,10 +53,19 @@ struct Vertex
 
 class Model
 {
+private:
+    std::string name;
 public:
     std::vector<Vertex> vertices;
     std::vector<uint32_t> indices;
+    Buffer vertexBuffer;
+    Buffer indexBuffer;
+    Model(std::string model);
     void loadModel();
+    void makeVertexBuffer();
+    void makeIndexBuffer();
+    void init();
+    void render(VkCommandBuffer commandbuffer);
 };
 
 #endif
